+++ src/main/java/hibernate/semina/generic/GenericService.java
@@ -3,9 +3,9 @@ package hibernate.semina.generic;
 import java.util.List;
 
 public interface GenericService<M extends Model<ID>, ID> {
-	public int count();
+	public long count();
 
-	public int count(M model);
+	public long count(M model);
 
 	public M create(M model);
